New algorithm - research input by letters
Fix code and add new feature which opportunity comparison two inputs.
Any input have some letters, then safe have determine code.
Then inject input must be the same for aim broke safe at all.

for example:
abcd and abod is not the same, but abbc and abbc is the same, so this way safe is broken.